Add common type aliases

+++ includes/Raytracer/TextureManager.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "Utils/Common.h"
 #include "TextureData.h"
 #include <glbinding/gl/gl.h>
 class TextureManager
@@ -6,7 +7,7 @@ class TextureManager
 public:
 	TextureManager() = delete;
 	TextureManager(const TextureManager&) = delete;
-	static unsigned int loadTextureFromData(TextureData & textureData, bool gamma);
+	static uint32 loadTextureFromData(TextureData & textureData, bool gamma);
 	static gl::GLenum getTextureFormatFromData(TextureData & textureData);
-	static gl::GLenum getTextureFormatFromData(int componentCount);
+	static gl::GLenum getTextureFormatFromData(uint8 componentCount);
 };
